+++ minigpu/CHANGELOG.md
@@ -1,3 +1,12 @@
+# minigpu
+
+## 1.0.1
+
+- breaking: Uses dart native assets
+see updated readme.
+- implements platform stub for native assets to coexist with flutter plugins.
+- uses native_toolchain_cmake 0.0.4
+
 ## 1.0.0
 
 - Initial version.